Introduce STATSD_TAGS to add fixed tags to all stats
This can be used to tag stats with instance-specific context like build version, whether it's a staging vs prod workload, etc.

A word of caution: statsd has no common standard regarding how tags are encoded.
The gostats library follows a conversion of <metric>.__<key1>=<value2>.__<key2>=<value2>
but there are other libraries that do different things. I create out
https://github.com/lyft/gostats/pull/99 attempting to add more flexibility to gostats, but
decided to abandon it for now.

Longer term, it'd probably be better to adopt a more standarized metrics library (maybe
OpenTelemetry if it's ready?...) and let people plug in exporters for statsd or prometheus
or whatever else they want to use.

Signed-off-by: David Weitzman <dweitzman@pinterest.com>